single function for all cross checks

+++ src/core/utils.ts
@@ -2,9 +2,9 @@
  * Calculates array sum
  *
  * @export
- * @param {Uint8ClampedArray} arr - Input array.
+ * @param {UInt16Array} arr - Input array.
  * @returns - The sum of the array.
  */
-export function sumArray(arr: Int32Array): number {
+export function sumArray(arr: Uint16Array): number {
     return arr.reduce((sum, x) => sum + x);
 }
